update-patchelf
Update patchelf to current git version

The current version fixes a bug which broke the ability to strip
binaries with debug symbols:
https://github.com/NixOS/patchelf/pull/117

With the former version of patchelf, I would get "not enough room for
program headers, try linking with -N" when trying to strip the binaries.

This patch brings patchelf in sync with the commit
27ffe8ae871e7a186018d66020ef3f6162c12c69 of patchelf's git
repository.

Also update patchelf's build process to match the logic in upstream's
automake files (thanks to Frédéric Bonnard for the patch to properly use
getconf PAGESIZE).

Bug: https://bugreports.qt.io/browse/PYSIDE-740
Last-Update: 2018-07-09


Gbp-Pq: Name update-patchelf.patch