[PATCH 1/3] Fix Python 3.8 problems
This patch fixes some refcounting problems with Python 3.8 .
One incompatible change was announced in the what's new
document, but actually there were two more problems which
were not explicitly mentioned but took much time to sort out.

The patch is compatible with the limited API changes
(tested with debug build and API error disabled).
It is also independent of the Python version which is
full Limited API support.

For more info, see the documentation mentioned below.

The flag error is circumvented now! We either find a better
solution or leave it as it is. For now this is ok.

Fixes: PYSIDE-939
Change-Id: Iff4a9816857a6ebe86efd4b654d8921e4e464939
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>

Gbp-Pq: Name 0002-Fix-Python-3.8-problems.patch